update booking and customer

+++ database/migrations/2025_12_14_073613_create_customer_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('phoneNumber');
             $table->string('college');
             $table->string('faculty');
-            $table->float('depoBalance', 8, 2);
+            $table->decimal('depoBalance', 8, 2);
             $table->boolean('isBlacklisted')->default(false);
             $table->string('blacklistReason')->nullable();
